+++ src/util/figure.ts
@@ -1,13 +1,19 @@
 ﻿import { index } from '../parser/inline/extension/label';
 import { isGroup } from '../parser/inline';
+import { parse } from '../parser/api';
 import { define } from 'typed-dom';
 
 export function figure(source: DocumentFragment | HTMLElement): void {
   let base = '0';
   const indexes = new Map<string, string>();
   const exclusions = new Set<Element>();
-  return void source.querySelectorAll<HTMLElement>('figure[data-label][data-group]')
+  return void source.querySelectorAll<HTMLElement>('figure[data-label][data-group], h2[id]')
     .forEach(fig => {
+      if (fig.matches('h2')) {
+        assert(fig.parentNode === source);
+        if (base === '0') return;
+        fig = parse(`[:$-${+base.split('.')[0] + 1}.0]\n$$\n$$`).querySelector('figure')!;
+      }
       if (fig.matches('.example figure')) return;
       if (fig.parentElement !== source && fig.parentElement instanceof HTMLQuoteElement) {
         return exclusions.has(fig.parentElement)
@@ -17,10 +23,14 @@ export function figure(source: DocumentFragment | HTMLElement): void {
       }
       const label = fig.getAttribute('data-label')!;
       const group = fig.getAttribute('data-group')!;
-      const idx = index(label, indexes.get(group) || base);
-      idx.endsWith('.0')
-        ? void indexes.clear() || (base = idx)
-        : void indexes.set(group, idx);
+      let idx = index(label, indexes.get(group) || base);
+      if (idx.endsWith('.0')) {
+        base = idx = idx.startsWith('0.')
+          ? `${(indexes.get(group) || base).split('.', 1)[0]}.${idx.slice(2)}`
+          : idx;
+        void indexes.clear();
+      }
+      void indexes.set(group, idx);
       void fig.setAttribute('data-index', idx);
       const figindex = fig.lastElementChild!.previousElementSibling!;
       assert(figindex.matches('.figindex'));